refactor: move provider errors to sigle file and normalize constructors
