+++ v-next/hardhat/src/internal/builtin-plugins/network-manager/http-provider.ts
@@ -215,7 +215,7 @@ export class HttpProvider extends BaseProvider {
         }
 
         // eslint-disable-next-line no-restricted-syntax -- allow throwing ProviderError
-        throw new LimitExceededError(e);
+        throw new LimitExceededError(undefined, e);
       }
 
       throw e;
